code-vault-ai/requirements.txt
sentence-transformers>=2.2.2
pandas
numpy
tiktoken
tqdm
python-dotenv 